IAEA_Semantic_Search/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0